+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/disapatchersevlet/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe.handleradapter;
+package webmvc.org.springframework.web.servlet.mvc.disapatchersevlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
