+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/disapatchersevlet/HandlerMapping.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe.handlermapping;
+package webmvc.org.springframework.web.servlet.mvc.disapatchersevlet;
 
 import jakarta.servlet.http.HttpServletRequest;
