+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -4,9 +4,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import webmvc.org.springframework.web.servlet.mvc.tobe.handlermapping.AnnotationHandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.handlermapping.ControllerHandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.handlermapping.HandlerExecution;
+import webmvc.org.springframework.web.servlet.mvc.asis.ControllerHandlerMapping;
 
 import java.lang.reflect.InvocationTargetException;
